# infrastructure/api-swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Appraisal Management API",
                 "version":  "1.0.0",
                 "description":  "Complete API for appraisal management platform"
             },
    "servers":  [
                    {
                        "url":  "",
                        "description":  "Backend URL set by APIM serviceUrl"
                    }
                ],
    "paths":  {
...
                  "/api/orders/{orderId}":  {
                                                "get":  {
                                                            "operationId":  "getOrder",
                                                            "summary":  "Get order by ID",
                                                            "tags":  [
                                                                         "Orders"
                                                                     ],
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "orderId",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Order retrieved"
                                                                                      }
                                                                          }
                                                        }
                                            },
...
              }
}